Move state to its own module Co-authored-by: Brian Leung <leungbk@posteo.net>
